Added functionality to admin pages
Made nav-data use proper links
Made animallist page functional
Fixed missing method for deleteDog (Backend)

+++ Backend/src/main/java/com/it192/backend/service/DogService.java
@@ -44,6 +44,8 @@ public class DogService implements IDogService {
 
     @Override
     public void deleteDog(long id) {
-        dogRepository.deleteById(id);
+        Optional<Dog> dog = dogRepository.findById(id);
+
+        if(dog.isPresent()) dogRepository.deleteById(id);
     }
 }
